% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlunit-package.R
\docType{package}
\name{htmlunit}
\alias{htmlunit}
\title{Tools to Scrape Dynamic Web Content via the 'HtmlUnit' Java Library}
\description{
\code{HtmlUnit} (\url{http://htmlunit.sourceforge.net/}) is \emph{a "'GUI'-Less
browser for 'Java' programs". It models 'HTML' documents and provides an 'API'
that allows one to invoke pages, fill out forms, click links and more just like
one does in a "normal" browser. The library has fairly good and constantly
improving 'JavaScript' support and is able to work even with quite complex 'AJAX'
libraries, simulating 'Chrome', 'Firefox' or 'Internet Explorer' depending on
the configuration used. It is typically used for testing purposes or to retrieve
information from web sites.}
}
\details{
Tools are provided to work with this library at a higher level than provided by
the exposed 'Java' libraries in the \href{https://gitlab.com/hrbrmstr/htmlunitjars}{\code{htmlunitjars}}
package.
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/htmlunit}
\item BugReports: \url{https://gitlab.com/hrbrmstr/htmlunit/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
