% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-client.R
\name{print.browserinfo}
\alias{print.browserinfo}
\title{Print method for \code{browserinfo} objects}
\usage{
\method{print}{browserinfo}(x, ...)
}
\arguments{
\item{x}{\code{browserinfo} object}

\item{...}{unused}
}
\value{
\code{x}
}
\description{
Print method for \code{browserinfo} objects
}
\keyword{internal}
