% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-client.R
\name{print.webclient}
\alias{print.webclient}
\title{Print method for \code{webclient} objects}
\usage{
\method{print}{webclient}(x, ...)
}
\arguments{
\item{x}{\code{webclient} object}

\item{...}{unused}
}
\value{
\code{x}
}
\description{
Print method for \code{webclient} objects
}
\keyword{internal}
