% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc-click.R
\name{wc_click_on}
\alias{wc_click_on}
\title{Click on a DOM element in a webclient loaded page}
\usage{
wc_click_on(wc_obj, css, xpath)
}
\arguments{
\item{wc_obj}{a \code{webclient} object}

\item{css, xpath}{Node to click on. Supply one of css or xpath depending on whether you want to use a css or xpath 1.0 selector.}
}
\description{
Click on a DOM element in a webclient loaded page
}
\note{
The caller does not have to assign the output of this function to a
variable as the browser state is managed internally by HtmlUnit.
}
