% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc-status.R
\name{wc_content_length}
\alias{wc_content_length}
\title{Return content length of the last web request for current page}
\usage{
wc_content_length(wc_obj)
}
\arguments{
\item{wc_obj}{a \code{webclient} object}
}
\value{
the content length (in bytes) of the web request or \code{NULL} if no active page
}
\description{
Return content length of the last web request for current page
}
\note{
This is an information retrieval function that does not return
the \code{wc_obj} so must be the last function call in a \code{webclient} pipe.
}
