% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc-options.R
\name{wc_css}
\alias{wc_css}
\title{Enable/Disable CSS support}
\usage{
wc_css(wc_obj, enable)
}
\arguments{
\item{wc_obj}{a \code{webclient} object}

\item{enable}{if \code{TRUE} enable CSS support (which is the HtmlUnit default)}
}
\value{
the \code{webclient} object (invisibly)
}
\description{
Enable/Disable CSS support
}
\note{
The caller does not have to assign the output of this function to a
variable as the browser state is managed internally by HtmlUnit.
}
\seealso{
Other wc_opts: 
\code{\link{wc_dnt}()},
\code{\link{wc_geo}()},
\code{\link{wc_img_dl}()},
\code{\link{wc_resize}()},
\code{\link{wc_timeout}()},
\code{\link{wc_use_insecure_ssl}()},
\code{\link{wc_wait}()}
}
\concept{wc_opts}
