% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-client.R
\name{wc_go}
\alias{wc_go}
\title{Visit a URL}
\usage{
wc_go(wc_obj, url)
}
\arguments{
\item{wc_obj}{a \code{webclient} object}

\item{url}{URL to retrieve}
}
\value{
the \code{webclient} object (invisibly)
}
\description{
Visit a URL
}
\note{
The caller does not have to assign the output of this function to a
variable as the browser state is managed internally by HtmlUnit.
}
