% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc-status.R
\name{wc_headers}
\alias{wc_headers}
\title{Return response headers of the last web request for current page}
\usage{
wc_headers(wc_obj)
}
\arguments{
\item{wc_obj}{a \code{webclient} object}
}
\value{
the response headers of the web request as a data frame or \code{NULL} if
no active page
}
\description{
Return response headers of the last web request for current page
}
\note{
This is an information retrieval function that does not return
the \code{wc_obj} so must be the last function call in a \code{webclient} pipe.
}
