% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc-html-nodes.R
\name{wc_html_nodes}
\alias{wc_html_nodes}
\title{Select nodes from web client active page html content}
\usage{
wc_html_nodes(wc_obj, css, xpath)
}
\arguments{
\item{wc_obj}{a \code{webclient} object}

\item{css, xpath}{Nodes to select. Supply one of css or xpath depending on whether you want to use a css or xpath 1.0 selector.}
}
\description{
Select nodes from web client active page html content
}
\examples{
\dontrun{
wc <- web_client()

wc \%>\% wc_go("https://usa.gov/")

wc \%>\%
  wc_html_nodes("a") \%>\%
  sapply(wc_html_text)

wc \%>\%
  wc_html_nodes(xpath=".//a") \%>\%
  sapply(wc_html_text)

wc \%>\%
  wc_html_nodes(xpath=".//a") \%>\%
  sapply(wc_html_attr, "href")
}
}
