% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc-html-nodes.R
\name{wc_html_text}
\alias{wc_html_text}
\alias{wc_html_attr}
\alias{wc_html_name}
\title{Extract attributes, text and tag name from webclient page html content}
\usage{
wc_html_text(dom_node, trim = FALSE)

wc_html_attr(dom_node, attr)

wc_html_name(dom_node)
}
\arguments{
\item{dom_node}{a webclient page DOM node (likely produced by \code{\link[=wc_html_nodes]{wc_html_nodes()}})}

\item{trim}{if \code{TRUE} will trim leading/trailing white space}

\item{attr}{name of attribute to retrieve}
}
\description{
Extract attributes, text and tag name from webclient page html content
}
\examples{
\dontrun{
wc <- web_client()

wc \%>\% wc_go("https://usa.gov/")

wc \%>\%
  wc_html_nodes("a") \%>\%
  sapply(wc_html_text)

wc \%>\%
  wc_html_nodes(xpath=".//a") \%>\%
  sapply(wc_html_text)

wc \%>\%
  wc_html_nodes(xpath=".//a") \%>\%
  sapply(wc_html_attr, "href")
}
}
