% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc-inspect.R
\name{wc_inspect}
\alias{wc_inspect}
\title{Perform a "Developer Tools"-like Network Inspection of a URL}
\usage{
wc_inspect(url, js_delay = 5000L)
}
\arguments{
\item{url}{URL to fetch}

\item{js_delay}{(ms) How long to wait for JavaScript to execute/XHRs to load? (Default: 5000)}
}
\description{
Retrieves \emph{all} content loaded
}
