% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc-options.R
\name{wc_resize}
\alias{wc_resize}
\title{Resize the virtual browser window}
\usage{
wc_resize(wc_obj, h, w)
}
\arguments{
\item{wc_obj}{a \code{webclient} object}

\item{h, w}{height and width (pixels)}
}
\value{
the \code{webclient} object (invisibly)
}
\description{
Resize the virtual browser window
}
\note{
The caller does not have to assign the output of this function to a
variable as the browser state is managed internally by HtmlUnit.
}
\seealso{
Other wc_opts: \code{\link{wc_css}}, \code{\link{wc_dnt}},
  \code{\link{wc_geo}}, \code{\link{wc_img_dl}},
  \code{\link{wc_timeout}},
  \code{\link{wc_use_insecure_ssl}}, \code{\link{wc_wait}}
}
\concept{wc_opts}
