% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc-status.R
\name{wc_status}
\alias{wc_status}
\title{Return status code of web request for current page}
\usage{
wc_status(wc_obj, url)

wc_status(wc_obj, url)
}
\arguments{
\item{wc_obj}{a `webclient` object}

\item{wc_obj}{a `webclient` object}
}
\value{
the HTTP status code and message of the web request or `NULL` if no active page

the content type o the web request or `NULL` if no active page
}
\description{
Return status code of web request for current page

Return content type of web request for current page
}
\note{
This is an information retrieval function that does not return
      the `wc_obj` so must be the last function call in a `webclient` pipe.

This is an information retrieval function that does not return
      the `wc_obj` so must be the last function call in a `webclient` pipe.
}
