% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc-options.R
\name{wc_timeout}
\alias{wc_timeout}
\title{Change default request timeout}
\usage{
wc_timeout(wc_obj, timeout)
}
\arguments{
\item{wc_obj}{a `webclient` object}

\item{timneout}{timeout (ms); The timeout is used twice. The first is for making
the socket connection, the second is for data retrieval. If the
time is critical you must allow for twice the time specified here.}
}
\value{
the `webclient` object (invisibly)
}
\description{
Change default request timeout
}
\note{
The caller does not have to assign the output of this function to a
      variable as the browser state is managed internally by HtmlUnit.
}
\seealso{
Other wc_opts: \code{\link{wc_css}}, \code{\link{wc_dnt}},
  \code{\link{wc_geo}}, \code{\link{wc_img_dl}},
  \code{\link{wc_resize}},
  \code{\link{wc_use_insecure_ssl}}, \code{\link{wc_wait}}
}
\concept{wc_opts}
