% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc-status.R
\name{wc_title}
\alias{wc_title}
\title{Return page title for current page}
\usage{
wc_title(wc_obj)
}
\arguments{
\item{wc_obj}{a `webclient` object}
}
\value{
page title of the current page `NULL` if no active page
}
\description{
Return page title for current page
}
\note{
This is an information retrieval function that does not return
      the `wc_obj` so must be the last function call in a `webclient` pipe.
}
