% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc-options.R
\name{wc_use_insecure_ssl}
\alias{wc_use_insecure_ssl}
\title{Enable/Disable Ignoring SSL Validation Issues}
\usage{
wc_use_insecure_ssl(wc_obj, enable)
}
\arguments{
\item{wc_obj}{a \code{webclient} object}

\item{enable}{if \code{TRUE} the client will accept connections to any host,
regardless of whether they have valid certificates or not}
}
\value{
the \code{webclient} object (invisibly)
}
\description{
Enable/Disable Ignoring SSL Validation Issues
}
\note{
The caller does not have to assign the output of this function to a
variable as the browser state is managed internally by HtmlUnit.
}
\seealso{
Other wc_opts: 
\code{\link{wc_css}()},
\code{\link{wc_dnt}()},
\code{\link{wc_geo}()},
\code{\link{wc_img_dl}()},
\code{\link{wc_resize}()},
\code{\link{wc_timeout}()},
\code{\link{wc_wait}()}
}
\concept{wc_opts}
