% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc-options.R
\name{wc_wait}
\alias{wc_wait}
\title{Block HtlUnit final rendering blocks until all background JavaScript tasks have finished executing}
\usage{
wc_wait(wc_obj, js_delay = 2000L)
}
\arguments{
\item{wc_obj}{a \code{webclient} object}

\item{js_delay}{number of ms to wait/block}
}
\description{
Block HtlUnit final rendering blocks until all background JavaScript tasks have finished executing
}
\note{
The caller does not have to assign the output of this function to a
variable as the browser state is managed internally by HtmlUnit.
}
\seealso{
Other wc_opts: \code{\link{wc_css}}, \code{\link{wc_dnt}},
  \code{\link{wc_geo}}, \code{\link{wc_img_dl}},
  \code{\link{wc_resize}}, \code{\link{wc_timeout}},
  \code{\link{wc_use_insecure_ssl}}
}
\concept{wc_opts}
