% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-client.R
\name{web_client}
\alias{web_client}
\alias{webclient}
\title{Create a new HtmlUnit WebClient instance}
\usage{
web_client(
  emulate = c("best", "chrome", "firefox", "ie", "edge"),
  proxy_host = NULL,
  proxy_port = NULL
)

webclient(
  emulate = c("best", "chrome", "firefox", "ie", "edge"),
  proxy_host = NULL,
  proxy_port = NULL
)
}
\arguments{
\item{emulate}{browser to emulate; one of "\code{best}", "\code{chrome}", "\code{firefox}", "\code{ie}"}

\item{proxy_host, proxy_port}{the server/port that will act as proxy (default
\code{NULL} = no proxy)}
}
\value{
\code{webclient} object
}
\description{
A new HtmlUnit web client (virtual browser) will be created and a \code{webclient}
object will be returned.
}
\details{
This is part of the \code{htmlunit} DSL interface.s
}
\examples{
w <- web_client()
wc_browser_info(w)
}
\concept{dsl}
