% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlunit-package.R
\docType{package}
\name{htmlunit}
\alias{htmlunit}
\alias{htmlunit-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/htmlunit}
\item BugReports: \url{https://gitlab.com/hrbrmstr/htmlunit/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
