% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-client.R
\name{print.browserinfo}
\alias{print.browserinfo}
\title{Print method for `browserinfo` objects}
\usage{
\method{print}{browserinfo}(x, ...)
}
\arguments{
\item{x}{`browserinfo` object}

\item{...}{unused}
}
\description{
Print method for `browserinfo` objects
}
\keyword{internal}
