% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-client.R
\name{print.webclient}
\alias{print.webclient}
\title{Print method for `webclient` objects}
\usage{
\method{print}{webclient}(x, ...)
}
\arguments{
\item{x}{`webclient` object}

\item{...}{unused}
}
\description{
Print method for `webclient` objects
}
\keyword{internal}
