% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc-inspect.R
\name{wc_inspect}
\alias{wc_inspect}
\title{Perform a "Developer Tools"-like Network Inspection of a URL}
\usage{
wc_inspect(
  url,
  js_delay = 5000L,
  timeout = 30000L,
  css = FALSE,
  images = FALSE
)
}
\arguments{
\item{url}{URL to fetch}

\item{js_delay}{(ms) How long to wait for JavaScript to execute/XHRs to load? (Default: 5000)}

\item{timeout}{Sets the timeout (milliseconds) of the web connection. Set to zero for an infinite wait.
Defaults to \code{30000}. Note: The timeout is used twice. The first is for making the socket
connection, the second is for data retrieval. If the time is critical you must allow for twice
the time specified here.}

\item{css, images}{enable CSS/download images? (default \code{FALSE})}
}
\description{
Retrieves \emph{all} content loaded
}
