% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-client.R
\name{web_client}
\alias{web_client}
\title{Create a new HtmlUnit WebClient instance}
\usage{
web_client(emulate = c("best", "chrome", "firefox", "ie"),
  proxy_host = NULL, proxy_port = NULL)
}
\arguments{
\item{emulate}{browser to emulate; one of "`best`", "`chrome`", "`firefox`", "`ie`"}

\item{proxy_host, proxy_port}{the server/port that will act as proxy (default
`NULL` = no proxy)}
}
\value{
`webclient` object
}
\description{
A new HtmlUnit web client (virtual browser) will be created and a `webclient`
object will be returned.
}
\details{
This is part of the `htmlunit` DSL interface.s
}
\concept{dsl}
